requirements.txt
asgiref==3.3.1
boto3==1.17.13
certifi==2020.12.5
chardet==4.0.0
coverage==5.3.1
Django==3.1.7
django-crispy-forms==1.10.0
django-tinymce==3.2.0
django-storages==1.11.1
gunicorn==20.0.4
idna==2.10
imgurpython==1.1.7
Pillow==8.2.0
pip==21.0.1
setuptools==51.3.3
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
django-heroku==0.3.1
